IPCC color palette for Sea Surface Temperature

Default colors for Sea Surface Temperature, as indicated by the IPCC AR6 WGI style guide, when using a divergent color palette

Usage

palette_IPCC_temp()

Source

Page 8 of the 2019 ICPP Style guide: https://www.ipcc.ch/site/assets/uploads/2019/04/IPCC-visual-style-guide.pdf

Mark Harrower & Cynthia A. Brewer (2003) ColorBrewer.org: An Online Tool for Selecting Colour Schemes forMaps, The Cartographic Journal, 40:1, 27-37, DOI: 10.1179/000870403235002042

Value

A character vector containing 11 colors,

Examples

library(EVR628tools)
palette_IPCC_temp()
#>  [1] "#67001F" "#B1182B" "#D5604D" "#F3A481" "#FCDAC6" "#F6F6F6" "#D0E4EF"
#>  [8] "#91C4DD" "#4392C2" "#2166AB" "#053061"